Addition of the full selection database file
I am not able to commit this full selection file, because it is too large for GitHub.com, it is available through https://www.astrochemistry.org/pahdb/theoretical/3.20/download/send/complete/xml and selecting theory for the download setting. Renamed some of the files for convenience and removed the settings.json because it does not select a the environment kernel by default.